Migrate the `Files.readString` interceptor to the NIO interceptor declarations
